dog-breed/requirements.txt
torch
tf-keras
datasets
evaluate
transformers[sentencepiece]
Pillow
